+++ mirrord-config/src/feature.rs
@@ -1,0 +1,24 @@
+use mirrord_config_derive::MirrordConfig;
+use serde::Deserialize;
+
+use crate::{
+    config::source::MirrordConfigSource, env::EnvFileConfig, fs::FsConfig,
+    network::NetworkFileConfig, util::ToggleableConfig,
+};
+
+#[derive(MirrordConfig, Deserialize, Default, PartialEq, Eq, Clone, Debug)]
+#[serde(deny_unknown_fields)]
+#[config(map_to = FeatureConfig)]
+pub struct FeatureFileConfig {
+    #[serde(default)]
+    #[config(nested)]
+    pub env: ToggleableConfig<EnvFileConfig>,
+
+    #[serde(default)]
+    #[config(nested)]
+    pub fs: ToggleableConfig<FsConfig>,
+
+    #[serde(default)]
+    #[config(nested)]
+    pub network: ToggleableConfig<NetworkFileConfig>,
+}