+++ tests/src/sanity.rs
@@ -187,7 +187,7 @@ mod tests {
     impl Application {
         async fn run(
             &self,
-            pod_name: &str,
+            target: &str,
             namespace: Option<&str>,
             args: Option<Vec<&str>>,
         ) -> TestProcess {
@@ -208,7 +208,7 @@ mod tests {
                 Application::Go18HTTP => vec!["go-e2e/18"],
                 Application::Go19HTTP => vec!["go-e2e/19"],
             };
-            run(process_cmd, pod_name, namespace, args).await
+            run(process_cmd, target, namespace, args).await
         }
 
         fn assert(&self, process: &TestProcess) {
@@ -256,7 +256,7 @@ mod tests {
 
     async fn run(
         process_cmd: Vec<&str>,
-        pod_name: &str,
+        target: &str,
         namespace: Option<&str>,
         args: Option<Vec<&str>>,
     ) -> TestProcess {
@@ -264,14 +264,14 @@ mod tests {
         let temp_dir = tempdir::TempDir::new("test").unwrap();
         let mut mirrord_args = vec![
             "exec",
-            "--pod-name",
-            pod_name,
+            "--target",
+            target,
             "-c",
             "--extract-path",
             temp_dir.path().to_str().unwrap(),
         ];
         if let Some(namespace) = namespace {
-            mirrord_args.extend(["--pod-namespace", namespace].into_iter());
+            mirrord_args.extend(["--target-namespace", namespace].into_iter());
         }
         if let Some(args) = args {
             mirrord_args.extend(args.into_iter());
@@ -288,7 +288,6 @@ mod tests {
         env.insert("MIRRORD_AGENT_IMAGE", "test");
         env.insert("MIRRORD_CHECK_VERSION", "false");
         env.insert("MIRRORD_AGENT_RUST_LOG", "warn,mirrord=debug");
-        env.insert("MIRRORD_IMPERSONATED_CONTAINER_NAME", CONTAINER_NAME);
         env.insert("MIRRORD_AGENT_COMMUNICATION_TIMEOUT", "180");
         env.insert("RUST_LOG", "warn,mirrord=debug");
         let server = Command::new(path)
@@ -367,7 +366,7 @@ mod tests {
     pub struct KubeService {
         name: String,
         namespace: String,
-        pod_name: String,
+        target: String,
         _pod: ResourceGuard,
         _service: ResourceGuard,
     }
@@ -475,18 +474,18 @@ mod tests {
         let service_guard = ResourceGuard::create(&service_api, name.to_string(), &service).await;
         watch_resource_exists(&service_api, "default").await;
 
-        let pod_name = get_pod_instance(&kube_client, &name, namespace)
+        let target = get_pod_instance(&kube_client, &name, namespace)
             .await
             .unwrap();
         let pod_api: Api<Pod> = Api::namespaced(kube_client.clone(), namespace);
-        await_condition(pod_api, &pod_name, is_pod_running())
+        await_condition(pod_api, &target, is_pod_running())
             .await
             .unwrap();
 
         KubeService {
             name: name.to_string(),
             namespace: namespace.to_string(),
-            pod_name,
+            target: format!("pod/{}/container/{}", target, CONTAINER_NAME),
             _pod: pod_guard,
             _service: service_guard,
         }
@@ -633,7 +632,7 @@ mod tests {
         let kube_client = kube_client.await;
         let url = get_service_url(kube_client.clone(), &service).await;
         let mut process = application
-            .run(&service.pod_name, Some(&service.namespace), agent.flag())
+            .run(&service.target, Some(&service.namespace), agent.flag())
             .await;
         process.wait_for_line(Duration::from_secs(120), "daemon subscribed");
         send_requests(&url, false).await;
@@ -670,7 +669,7 @@ mod tests {
         let kube_client = kube_client.await;
         let url = get_service_url(kube_client.clone(), &service).await;
         let mut process = application
-            .run(&service.pod_name, Some(&service.namespace), agent.flag())
+            .run(&service.target, Some(&service.namespace), agent.flag())
             .await;
         process.wait_for_line(Duration::from_secs(300), "daemon subscribed");
         send_requests(&url, false).await;
@@ -710,7 +709,7 @@ mod tests {
 
         let mut process = run(
             command,
-            &service.pod_name,
+            &service.target,
             Some(&service.namespace),
             Some(args),
         )
@@ -737,7 +736,7 @@ mod tests {
         let args = vec!["--rw"];
         let mut process = run(
             python_command,
-            &service.pod_name,
+            &service.target,
             Some(&service.namespace),
             Some(args),
         )
@@ -769,7 +768,7 @@ mod tests {
 
         let mut process = run(
             python_command,
-            &service.pod_name,
+            &service.target,
             Some(&service.namespace),
             None,
         )
@@ -789,7 +788,7 @@ mod tests {
             "node-e2e/remote_env/test_remote_env_vars_exclude_works.mjs",
         ];
         let mirrord_args = vec!["-x", "MIRRORD_FAKE_VAR_FIRST"];
-        let mut process = run(node_command, &service.pod_name, None, Some(mirrord_args)).await;
+        let mut process = run(node_command, &service.target, None, Some(mirrord_args)).await;
 
         let res = process.child.wait().await.unwrap();
         assert!(res.success());
@@ -806,7 +805,7 @@ mod tests {
             "node-e2e/remote_env/test_remote_env_vars_include_works.mjs",
         ];
         let mirrord_args = vec!["-s", "MIRRORD_FAKE_VAR_FIRST"];
-        let mut process = run(node_command, &service.pod_name, None, Some(mirrord_args)).await;
+        let mut process = run(node_command, &service.target, None, Some(mirrord_args)).await;
 
         let res = process.child.wait().await.unwrap();
         assert!(res.success());
@@ -822,7 +821,7 @@ mod tests {
             "node",
             "node-e2e/remote_dns/test_remote_dns_enabled_works.mjs",
         ];
-        let mut process = run(node_command, &service.pod_name, None, None).await;
+        let mut process = run(node_command, &service.target, None, None).await;
 
         let res = process.child.wait().await.unwrap();
         assert!(res.success());
@@ -838,7 +837,7 @@ mod tests {
             "node",
             "node-e2e/remote_dns/test_remote_dns_lookup_google.mjs",
         ];
-        let mut process = run(node_command, &service.pod_name, None, None).await;
+        let mut process = run(node_command, &service.target, None, None).await;
 
         let res = process.child.wait().await.unwrap();
         assert!(res.success());
@@ -866,7 +865,7 @@ mod tests {
         let mut flags = vec!["--steal"];
         agent.flag().map(|flag| flags.extend(flag));
         let mut process = application
-            .run(&service.pod_name, Some(&service.namespace), Some(flags))
+            .run(&service.target, Some(&service.namespace), Some(flags))
             .await;
 
         process.wait_for_line(Duration::from_secs(30), "daemon subscribed");
@@ -886,7 +885,7 @@ mod tests {
     pub async fn test_bash_remote_env_vars_works(#[future] service: KubeService) {
         let service = service.await;
         let bash_command = vec!["bash", "bash-e2e/env.sh"];
-        let mut process = run(bash_command, &service.pod_name, None, None).await;
+        let mut process = run(bash_command, &service.target, None, None).await;
 
         let res = process.child.wait().await.unwrap();
         assert!(res.success());
@@ -901,7 +900,7 @@ mod tests {
         let service = service.await;
         let bash_command = vec!["bash", "bash-e2e/env.sh", "exclude"];
         let mirrord_args = vec!["-x", "MIRRORD_FAKE_VAR_FIRST"];
-        let mut process = run(bash_command, &service.pod_name, None, Some(mirrord_args)).await;
+        let mut process = run(bash_command, &service.target, None, Some(mirrord_args)).await;
 
         let res = process.child.wait().await.unwrap();
         assert!(res.success());
@@ -916,7 +915,7 @@ mod tests {
         let service = service.await;
         let bash_command = vec!["bash", "bash-e2e/env.sh", "include"];
         let mirrord_args = vec!["-s", "MIRRORD_FAKE_VAR_FIRST"];
-        let mut process = run(bash_command, &service.pod_name, None, Some(mirrord_args)).await;
+        let mut process = run(bash_command, &service.target, None, Some(mirrord_args)).await;
 
         let res = process.child.wait().await.unwrap();
         assert!(res.success());
@@ -933,7 +932,7 @@ mod tests {
     pub async fn test_bash_file_exists(#[future] service: KubeService) {
         let service = service.await;
         let bash_command = vec!["bash", "bash-e2e/file.sh", "exists"];
-        let mut process = run(bash_command, &service.pod_name, None, None).await;
+        let mut process = run(bash_command, &service.target, None, None).await;
 
         let res = process.child.wait().await.unwrap();
         assert!(res.success());
@@ -951,7 +950,7 @@ mod tests {
     pub async fn test_bash_file_read(#[future] service: KubeService) {
         let service = service.await;
         let bash_command = vec!["bash", "bash-e2e/file.sh", "read"];
-        let mut process = run(bash_command, &service.pod_name, None, None).await;
+        let mut process = run(bash_command, &service.target, None, None).await;
 
         let res = process.child.wait().await.unwrap();
         assert!(res.success());
@@ -967,7 +966,7 @@ mod tests {
         let service = service.await;
         let bash_command = vec!["bash", "bash-e2e/file.sh", "write"];
         let args = vec!["--rw"];
-        let mut process = run(bash_command, &service.pod_name, None, Some(args)).await;
+        let mut process = run(bash_command, &service.target, None, Some(args)).await;
 
         let res = process.child.wait().await.unwrap();
         assert!(res.success());
@@ -979,7 +978,7 @@ mod tests {
     pub async fn test_go18_remote_env_vars_works(#[future] service: KubeService) {
         let service = service.await;
         let command = vec!["go-e2e-env/18"];
-        let mut process = run(command, &service.pod_name, None, None).await;
+        let mut process = run(command, &service.target, None, None).await;
         let res = process.child.wait().await.unwrap();
         assert!(res.success());
         process.assert_stderr();
@@ -990,7 +989,7 @@ mod tests {
     pub async fn test_go19_remote_env_vars_works(#[future] service: KubeService) {
         let service = service.await;
         let command = vec!["go-e2e-env/19"];
-        let mut process = run(command, &service.pod_name, None, None).await;
+        let mut process = run(command, &service.target, None, None).await;
         let res = process.child.wait().await.unwrap();
         assert!(res.success());
         process.assert_stderr();
@@ -1007,7 +1006,7 @@ mod tests {
             "node",
             "node-e2e/outgoing/test_outgoing_traffic_single_request.mjs",
         ];
-        let mut process = run(node_command, &service.pod_name, None, None).await;
+        let mut process = run(node_command, &service.target, None, None).await;
 
         let res = process.child.wait().await.unwrap();
         assert!(res.success());
@@ -1023,7 +1022,7 @@ mod tests {
             "node",
             "node-e2e/outgoing/test_outgoing_traffic_single_request_ipv6.mjs",
         ];
-        let mut process = run(node_command, &service.pod_name, None, None).await;
+        let mut process = run(node_command, &service.target, None, None).await;
 
         let res = process.child.wait().await.unwrap();
         assert!(res.success());
@@ -1039,7 +1038,7 @@ mod tests {
             "node-e2e/outgoing/test_outgoing_traffic_single_request.mjs",
         ];
         let mirrord_args = vec!["--no-outgoing"];
-        let mut process = run(node_command, &service.pod_name, None, Some(mirrord_args)).await;
+        let mut process = run(node_command, &service.target, None, Some(mirrord_args)).await;
 
         let res = process.child.wait().await.unwrap();
         assert!(res.success());
@@ -1054,7 +1053,7 @@ mod tests {
             "node",
             "node-e2e/outgoing/test_outgoing_traffic_many_requests.mjs",
         ];
-        let mut process = run(node_command, &service.pod_name, None, None).await;
+        let mut process = run(node_command, &service.target, None, None).await;
 
         let res = process.child.wait().await.unwrap();
         assert!(res.success());
@@ -1070,7 +1069,7 @@ mod tests {
             "node-e2e/outgoing/test_outgoing_traffic_many_requests.mjs",
         ];
         let mirrord_args = vec!["--no-outgoing"];
-        let mut process = run(node_command, &service.pod_name, None, Some(mirrord_args)).await;
+        let mut process = run(node_command, &service.target, None, Some(mirrord_args)).await;
 
         let res = process.child.wait().await.unwrap();
         assert!(res.success());
@@ -1085,7 +1084,7 @@ mod tests {
             "node",
             "node-e2e/outgoing/test_outgoing_traffic_make_request_after_listen.mjs",
         ];
-        let mut process = run(node_command, &service.pod_name, None, None).await;
+        let mut process = run(node_command, &service.target, None, None).await;
         let res = process.child.wait().await.unwrap();
         assert!(res.success());
         process.assert_stderr();
@@ -1099,7 +1098,7 @@ mod tests {
             "node",
             "node-e2e/outgoing/test_outgoing_traffic_make_request_localhost.mjs",
         ];
-        let mut process = run(node_command, &service.pod_name, None, None).await;
+        let mut process = run(node_command, &service.target, None, None).await;
         let res = process.child.wait().await.unwrap();
         assert!(res.success());
         process.assert_stderr();
@@ -1145,20 +1144,21 @@ mod tests {
         let mirrord_no_outgoing = vec!["--no-outgoing"];
         let mut process = run(
             node_command.clone(),
-            &target_service.pod_name,
+            &target_service.target,
             Some(&target_service.namespace),
             Some(mirrord_no_outgoing),
         )
         .await;
         let res = process.child.wait().await.unwrap();
         assert!(!res.success()); // Should fail because local process cannot reach service.
-        let logs = pod_api.logs(&internal_service.pod_name, &lp).await;
+        let stripped_target = internal_service.target.split('/').collect::<Vec<&str>>()[1];
+        let logs = pod_api.logs(stripped_target, &lp).await;
         assert_eq!(logs.unwrap(), "");
 
         // Run mirrord with outgoing enabled.
         let mut process = run(
             node_command,
-            &target_service.pod_name,
+            &target_service.target,
             Some(&target_service.namespace),
             None,
         )
@@ -1170,7 +1170,7 @@ mod tests {
         // Verify that the UDP message sent by the application reached the internal service.
         lp.follow = true; // Follow log stream.
         let logs = pod_api
-            .log_stream(&internal_service.pod_name, &lp)
+            .log_stream(stripped_target, &lp)
             .await
             .unwrap()
             .try_next()
@@ -1199,7 +1199,7 @@ mod tests {
             &port,
         ];
         let mirrord_args = vec!["--no-outgoing"];
-        let mut process = run(node_command, &service.pod_name, None, Some(mirrord_args)).await;
+        let mut process = run(node_command, &service.target, None, Some(mirrord_args)).await;
 
         // Listen for UDP message directly from application.
         let mut buf = [0; 27];
@@ -1214,7 +1214,7 @@ mod tests {
 
     pub async fn test_go(service: impl Future<Output = KubeService>, command: Vec<&str>) {
         let service = service.await;
-        let mut process = run(command, &service.pod_name, None, None).await;
+        let mut process = run(command, &service.target, None, None).await;
         let res = process.child.wait().await.unwrap();
         assert!(res.success());
         process.assert_stderr();